+++ src/chanjo2/main.py
@@ -1,15 +1,25 @@
 import logging
 import os
+from typing import List, Tuple
 
 import uvicorn
+from fastapi import FastAPI, status
+from fastapi.staticfiles import StaticFiles
+
 from chanjo2 import __version__
 from chanjo2.dbutil import engine
-from chanjo2.endpoints import cases, intervals, samples, coverage
+from chanjo2.endpoints import cases, intervals, samples, coverage, report
 from chanjo2.models.sql_models import Base
 from chanjo2.populate_demo import load_demo_data
-from fastapi import FastAPI, status
 
 LOG = logging.getLogger("uvicorn.access")
+APP_ROUTER_TAGS: List[Tuple] = [
+    (intervals.router, "intervals"),
+    (coverage.router, "coverage"),
+    (cases.router, "cases"),
+    (samples.router, "samples"),
+    (report.router, "report"),
+]
 
 
 def create_db_and_tables():
@@ -18,29 +28,21 @@ def create_db_and_tables():
 
 app = FastAPI()
 
-app.include_router(
-    intervals.router,
-    tags=["intervals"],
-    responses={status.HTTP_404_NOT_FOUND: {"description": "Not found"}},
-)
 
-app.include_router(
-    coverage.router,
-    tags=["coverage"],
-    responses={status.HTTP_404_NOT_FOUND: {"description": "Not found"}},
-)
+def configure_static(app):
+    """Configure static folder."""
+    exec_dir: str = os.path.dirname(__file__)
+    static_abs_file_path: str = os.path.join(exec_dir, "static/")
+    app.mount("/static", StaticFiles(directory=static_abs_file_path), name="static")
 
-app.include_router(
-    cases.router,
-    tags=["cases"],
-    responses={status.HTTP_404_NOT_FOUND: {"description": "Not found"}},
-)
 
-app.include_router(
-    samples.router,
-    tags=["samples"],
-    responses={status.HTTP_404_NOT_FOUND: {"description": "Not found"}},
-)
+# include app router tags
+for router, tag in APP_ROUTER_TAGS:
+    app.include_router(
+        router,
+        tags=[tag],
+        responses={status.HTTP_404_NOT_FOUND: {"description": "Not found"}},
+    )
 
 
 @app.on_event("startup")
@@ -56,6 +58,8 @@ async def on_startup():
 
     # Create database tables
     create_db_and_tables()
+
+    configure_static(app)
 
     if os.getenv("DEMO") or not os.getenv("MYSQL_DATABASE_NAME"):
         LOG.warning("Running a demo instance of Chanjo2")